Test works almost completely. Now its tobi's job to implement the proper business logic
https://github.com/metasfresh/metasfresh/issues/6732
