+++ e2e/cypress/fixtures/materialDisposition/create_material_dispo_masterdata_and_sales_order_spec.json
@@ -7,7 +7,6 @@
 
     "plantName" : "MFG Resource",
     "priceListName" : "PriceList",
-    "priceListVersionName" : "PriceListVersion",
 
     "componentProductName" : "BOM1-ComponentProduct",
